Modified and add single modal
