+++ practical04/src/components/NavBarComponent.vue
@@ -3,14 +3,6 @@
     <nav class="navbar navbar-light bg-success">
       <div class="container-fluid center">
         <span class="navbar-brand mb-0 h1">Car Showroom</span>
-        <button
-          type="button"
-          data-bs-toggle="modal"
-          data-bs-target="#staticBackdrop"
-          class="btn btn-primary"
-        >
-          Add Car
-        </button>
       </div>
     </nav>
   </div>
@@ -23,4 +15,4 @@ export default {
 </script>
 
 <!-- Add "scoped" attribute to limit CSS to this component only -->
-<style scoped></style>
+<style></style>
